backend/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.3

# Neo4j Graph Database
neo4j==5.25.0

# Redis
redis==5.1.1
redis-om==0.3.2

# Vector Database
qdrant-client==1.12.0

# AI/ML and RAG
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.2.3
langchain-community==0.3.7
langgraph==0.2.45
openai==1.54.4
anthropic==0.39.0
cohere==5.11.0
sentence-transformers==3.2.1
transformers==4.46.2
torch==2.5.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography==43.0.3
pyjwt==2.9.0
authlib==1.3.2

# FHIR
fhir.resources==7.1.0
fhirclient==4.2.1

# Blockchain (Hyperledger Fabric)
fabric-sdk-py==0.9.0

# Payment Processing
stripe==11.1.1

# Email & SMS
python-dotenv==1.0.1
emails==0.6.0
twilio==9.3.7

# File Processing
python-magic==0.4.27
pillow==11.0.0
pytesseract==0.3.13
pdf2image==1.17.0

# AWS SDK
boto3==1.35.60
botocore==1.35.60

# Data Validation
email-validator==2.2.0
phonenumbers==8.13.50

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# Monitoring & Logging
prometheus-client==0.21.0
sentry-sdk[fastapi]==2.17.0
python-json-logger==3.1.0

# Rate Limiting
slowapi==0.1.9

# Caching
aiocache==0.12.2

# Background Tasks
celery==5.4.0
celery[redis]==5.4.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.27.2
faker==30.8.2

# Code Quality
black==24.10.0
ruff==0.7.4
mypy==1.13.0
pre-commit==4.0.1

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.44

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
uuid==1.30
tenacity==9.0.0

